requirements.txt
RPi.GPIO
smbus2
pyserial
flask
flask-cors
